# Conflicts:
#	src/worker/src/main.cpp

+++ src/worker/src/params.h
@@ -3,21 +3,22 @@
 
 #include <array>
 #include <cstddef>
+#include <string>
 
 template <std::size_t Size> class Params
 {
   public:
-    Params(std::array<int, Size> &&start, std::array<int, Size> &&end, std::array<int, Size> &&step, std::string &&id);
-    std::array<int, Size> &get_current();
+    Params(std::array<float, Size> &&start, std::array<float, Size> &&end, std::array<float, Size> &&step, std::string &&id);
+    std::array<float, Size> &get_current();
     void next();
     int get_total_iterations();
     std::string get_id();
 
   private:
-    std::array<int, Size> start_;
-    std::array<int, Size> end_;
-    std::array<int, Size> step_;
-    std::array<int, Size> current_;
+    std::array<float, Size> start_;
+    std::array<float, Size> end_;
+    std::array<float, Size> step_;
+    std::array<float, Size> current_;
     int total_iterations;
     std::string id_;
 };
